move keybindings.py to data
